refactor: remove src barrel file
