+++ pages/learn/index.tsx
@@ -1,7 +1,7 @@
 import { GetStaticProps } from "next";
-import { getSortedLearnData } from "../../src/lib";
+import { getSortedLearnData } from "../../src/lib/learn.js";
 
-export { Learn as default } from "../../src";
+export { Learn as default } from "../../src/Learn/Learn.js";
 
 export const getStaticProps: GetStaticProps = async () => {
   const allLearnData = getSortedLearnData();
