+++ pages/posts/[id].tsx
@@ -1,7 +1,7 @@
 import { GetStaticPaths, GetStaticProps } from "next";
-import { getAllPostIds, getPostData } from "../../src";
+import { getAllPostIds, getPostData } from "../../src/lib/posts.js";
 
-export { PostContent as default } from "../../src";
+export { PostContent as default } from "../../src/PostContent/PostContent.js";
 
 export const getStaticPaths: GetStaticPaths = async () => {
   const paths = getAllPostIds();
