+++ pages/posts/index.tsx
@@ -1,7 +1,7 @@
 import { GetStaticProps } from "next";
-import { getSortedPostsData } from "../../src";
+import { getSortedPostsData } from "../../src/lib/posts.js";
 
-export { Posts as default } from "../../src";
+export { Posts as default } from "../../src/Posts/Posts.js";
 
 export const getStaticProps: GetStaticProps = async () => {
   const allPostsData = getSortedPostsData();
