--- src/LearnPage/LearnPage.tsx
@@ -1,3 +1,0 @@
-export function LearnPage(): JSX.Element {
-  return <>Learn | .NET</>;
-}
