--- src/LearnPage/index.ts
@@ -1,1 +1,0 @@
-export { LearnPage } from "./LearnPage";
