--- src/index.ts
@@ -1,7 +1,0 @@
-export { config } from "./config";
-export { Home } from "./Home/Home";
-export { getAllPostIds, getPostData, getSortedPostsData } from "./lib";
-export { PostContent } from "./PostContent/PostContent";
-export { Posts } from "./Posts";
-export { Learn } from "./Learn";
-export { LearnPage } from "./LearnPage";
